refactor(Custom Annotations): refactor dir, names e create readme
