+++ README.md
@@ -1,2 +1,26 @@
 # java-custom-annotations
-Proof of Concept to develop custom annotations in the Spring Boot ecosystem
+
+Este é um projeto de prova de conceito para desenvolver anotações personalizadas no ecossistema Spring Boot.
+
+## Requisitos
+
+- Java 17
+- Maven
+
+## Como executar
+
+1. Clone o repositório
+2. Navegue até a pasta do projeto
+3. Execute o comando `mvn spring-boot:run`
+
+## Funcionalidades
+
+Este projeto demonstra como criar e usar anotações personalizadas no Spring Boot.
+
+## Contribuição
+
+Pull requests são bem-vindos. Para mudanças importantes, por favor, abra uma issue primeiro para discutir o que você gostaria de mudar.
+
+## Licença
+
+[MIT](https://choosealicense.com/licenses/mit/)
